requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
sqlalchemy
psycopg2-binary
markdown
datetime